feat(api): Move submission email across to Resend
